ytmanager/pyproject.toml
[project]
name = "ytmanager"
version = "0.1.0"
description = "YouTube channel management system with analytics and auto-optimization"
requires-python = ">=3.11,<3.13"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "google-auth>=2.0",
    "google-auth-oauthlib>=1.0",
    "google-api-python-client>=2.0",
    "python-dotenv>=1.0",
    "aim>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-mock>=3.0",
    "ruff>=0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "unit: Unit tests (fast, no external deps)",
    "integration: Integration tests (mocked APIs)",
]
